+++ src/component/common/Header/Header.module.scss
@@ -32,165 +32,6 @@
       width: auto;
     }
   }
-
-  &__logo {
-    padding: 18px 35px 18px 0;
-
-    & > svg {
-      width: 80px;
-      height: 44px;
-
-      @include media.media-breakpoint-down(mobile) {
-        width: 58px;
-        height: 32px;
-      }
-    }
-  }
-
-  &__mega-menu {
-    height: 100%;
-    margin-right: auto;
-  }
-
-  &__auth-menu {
-    display: flex;
-    align-items: center;
-    list-style: none;
-  }
-
-  &__auth-link {
-    font-weight: normal;
-    float: right;
-    text-align: right;
-    line-height: 1.35;
-    font-family: "Noto Sans CJK KR", sans-serif;
-    font-size: 15px;
-    color: #a0d2f6;
-    padding: 0 20px;
-    margin: 30px 0;
-    cursor: pointer;
-
-    &:hover {
-      font-weight: 800;
-      color: #fff;
-    }
-
-    & > a {
-      color: inherit;
-      text-decoration: none;
-    }
-
-    & > button {
-      color: inherit;
-      text-decoration: none;
-      font-family: "Noto Sans CJK KR", sans-serif;
-      font-size: 15px;
-      background-color: #175c8e;
-      border: none;
-      cursor: pointer;
-
-      &:hover {
-        font-weight: 800;
-        color: #fff;
-      }
-    }
-  }
-}
-
-.megamenu {
-  &__panel {
-    display: block;
-    top: 80px;
-    overflow: hidden;
-    position: absolute;
-    left: 0;
-    margin: 0 auto;
-    width: 100%;
-    height: 0;
-    min-width: 1132px;
-    z-index: 4;
-    background: #15507a;
-    transition: height 0.5s;
-  }
-
-  &:hover,
-  &:focus-within {
-    & .megamenu__panel {
-      height: 75px;
-      box-sizing: border-box;
-    }
-  }
-
-  &__trigger-list {
-    list-style: none;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    flex: 1;
-    height: 100%;
-  }
-
-  &__trigger {
-    display: block;
-    border: none;
-    float: left;
-    padding: 0 20px;
-    height: 80px;
-    font-weight: 700;
-    font-family: "Noto Sans CJK KR", sans-serif;
-    font-size: 17px;
-    text-align: left;
-    line-height: 1;
-    color: #fff;
-    cursor: pointer;
-    background-color: #175c8e;
-
-    & > span {
-      padding: 31px 0 30px;
-    }
-
-    &:hover,
-    &:focus {
-      & > span {
-        padding-bottom: 28px;
-        border-bottom: 2px solid #fff;
-      }
-    }
-  }
-
-  &__content {
-    display: flex;
-    overflow: hidden;
-    width: calc(1132px - 135px);
-    height: 100%;
-    z-index: 5;
-    padding-left: 135px;
-    margin: 0 auto;
-    transition: height 0.2s, -webkit-transform 0.5s;
-    transition: height 0.2s, transform 0.5s;
-    list-style: none;
-  }
-
-  &__menu {
-    padding: 26px 0;
-    margin-right: 38px;
-  }
-
-  &__link {
-    line-height: 2;
-    display: block;
-    font-family: "Noto Sans CJK KR", sans-serif;
-    font-size: 13px;
-    padding-bottom: 10px;
-    color: rgb(255 255 255);
-    text-decoration: none;
-    text-align: left;
-    font-weight: 400;
-
-    &:hover {
-      font-weight: 800;
-    }
-  }
 }
 
 .add-menu-header {
